requirements.txt
streamlit==1.29.0
pandas==2.1.4
selenium==4.16.0
beautifulsoup4==4.12.2
requests==2.31.0
googlesearch-python==1.2.3
webdriver-manager==4.0.1
openpyxl==3.1.2
lxml==4.9.3